requirements.txt
black==23.11.0
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
flake8==6.1.0